fix я устал

+++ sivkov.filipp/F0/cmd.cpp
@@ -330,7 +330,7 @@ namespace sivkov
       }
     }
 
-    std::map< std::string, int > wordCount;
+    std::map< std::string, size_t > wordCount;
 
     for (const std::string& name : dictionaryNames)
     {
